requirements.txt
requests==2.32.3
python-dotenv==1.0.1
ddt==1.7.2
pytz==2024.1